+++ src/main/java/org/zalando/nakadi/controller/EventStreamController.java
@@ -24,7 +24,6 @@ import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.PartitionStatistics;
 import org.zalando.nakadi.domain.Timeline;
-import org.zalando.nakadi.exceptions.IllegalScopeException;
 import org.zalando.nakadi.exceptions.InternalNakadiException;
 import org.zalando.nakadi.exceptions.InvalidCursorException;
 import org.zalando.nakadi.exceptions.NakadiException;
@@ -221,9 +220,6 @@ public class EventStreamController {
                     Collections.singletonList(eventTypeName))) {
                 final EventType eventType = eventTypeRepository.findByName(eventTypeName);
 
-                // TODO: deprecate and remove previous authorization strategy
-                client.checkScopes(eventType.getReadScopes());
-
                 authorizeStreamRead(eventTypeName);
 
                 // validate parameters
@@ -287,9 +283,6 @@ public class EventStreamController {
                 writeProblemResponse(response, outputStream, e.asProblem());
             } catch (final InvalidCursorException e) {
                 writeProblemResponse(response, outputStream, PRECONDITION_FAILED, e.getMessage());
-            } catch (final IllegalScopeException e) {
-                // TODO: deprecate and remove previous authorization strategy
-                writeProblemResponse(response, outputStream, FORBIDDEN, e.getMessage());
             } catch (final AccessDeniedException e) {
                 writeProblemResponse(response, outputStream, FORBIDDEN, e.explain());
             } catch (final Exception e) {