+++ src/main/java/org/zalando/nakadi/security/Client.java
@@ -1,9 +1,5 @@
 package org.zalando.nakadi.security;
 
-import org.zalando.nakadi.exceptions.IllegalScopeException;
-
-import java.util.Set;
-
 public abstract class Client {
 
     private final String clientId;
@@ -11,8 +7,6 @@ public abstract class Client {
     public Client(final String clientId) {
         this.clientId = clientId;
     }
-
-    public abstract void checkScopes(Set<String> allowedScopes) throws IllegalScopeException;
 
     public String getClientId() {
         return clientId;