+++ src/main/java/org/zalando/nakadi/security/ClientResolver.java
@@ -20,7 +20,6 @@ import java.util.Optional;
 import java.util.Set;
 
 import static org.zalando.nakadi.config.SecuritySettings.AuthMode.OFF;
-import static org.zalando.nakadi.util.FeatureToggleService.Feature.CHECK_APPLICATION_LEVEL_PERMISSIONS;
 
 @Component
 public class ClientResolver implements HandlerMethodArgumentResolver {
@@ -46,8 +45,7 @@ public class ClientResolver implements HandlerMethodArgumentResolver {
                                   final NativeWebRequest request,
                                   final WebDataBinderFactory binderFactory) throws Exception {
         final Optional<String> clientId = Optional.ofNullable(request.getUserPrincipal()).map(Principal::getName);
-        if (!featureToggleService.isFeatureEnabled(CHECK_APPLICATION_LEVEL_PERMISSIONS)
-                || clientId.filter(settings.getAdminClientId()::equals).isPresent()
+        if (clientId.filter(settings.getAdminClientId()::equals).isPresent()
                 || settings.getAuthMode() == OFF) {
             return new FullAccessClient(clientId.orElse(FULL_ACCESS_CLIENT_ID));
         }